Change vagrant setup to be more organized

+++ Vagrantfile
@@ -16,11 +16,12 @@ Vagrant.configure(2) do |config|
   config.vm.provision "shell", inline: <<-SHELL
     apt-get update
     apt-get install -y nginx
-    rm /etc/nginx/sites-enabled/default
-    rm /etc/nginx/sites-available/default
-    cd /vagrant
-    cp server.khe-ui-staff.conf /etc/nginx/sites-available
-    ln -s /etc/nginx/sites-available/server.khe-ui-staff.conf /etc/nginx/sites-enabled
+    if [ -f /etc/nginx/sites-enabled/default ]; then
+      rm /etc/nginx/sites-enabled/default
+      rm /etc/nginx/sites-available/default
+    fi
+    cp /vagrant/config/staff.khe.conf /etc/nginx/sites-available
+    ln -s /etc/nginx/sites-available/staff.khe.conf /etc/nginx/sites-enabled
     service nginx reload
     echo "---------------------------------------------"
     echo "| See application at http://localhost:5000/ |"
